Move server from main into own module
