+++ src/actions/mod.rs
@@ -1,5 +1,5 @@
 pub use self::action::*;
-use crate::{errors::AppError, AppState};
+use crate::{errors::AppError, server::AppState};
 use anyhow::Result;
 use axum::{
     extract::{Path as AxumPath, State},
