+++ src/frontend/mod.rs
@@ -1,4 +1,4 @@
-use crate::{actions::ActionMsg, errors::AppError, AppState};
+use crate::{actions::ActionMsg, errors::AppError, server::AppState};
 
 use std::collections::HashMap;
