+++ src/lib.rs
@@ -1,0 +1,7 @@
+pub mod actions;
+pub mod config;
+pub mod errors;
+pub mod frontend;
+pub mod kobo_config;
+pub mod screenshot;
+pub mod server;
